create and get order api added

+++ src/app/modules/product/product.controller.ts
@@ -1,12 +1,13 @@
 import { Request, Response } from 'express';
 import ProductZodSchema from './product.zod';
 import { ProductServices } from './product.service';
+import { ModelProduct } from './product.model';
 
 const addProduct = async (req: Request, res: Response) => {
   try {
     const productData = req.body;
-    const productParsedData = ProductZodSchema.parse(productData);
-    const result = await ProductServices.addProductToDB(productParsedData);
+    const parsedproductData = ProductZodSchema.parse(productData);
+    const result = await ProductServices.addProductToDB(parsedproductData);
 
     res.status(200).send({
       success: true,
@@ -24,8 +25,16 @@ const addProduct = async (req: Request, res: Response) => {
 
 const getProducts = async (req: Request, res: Response) => {
   try {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const searchTerm: any = req.query.searchTerm;
     const products = await ProductServices.getAllProductsFromDB(searchTerm);
+
+    if (products.length <= 0) {
+      return res.status(200).send({
+        success: false,
+        message: 'Could not find any products',
+      });
+    }
     res.status(200).send({
       success: true,
       message: 'Products fetched successfully',
@@ -44,6 +53,12 @@ const getSingleProduct = async (req: Request, res: Response) => {
   try {
     const { productId } = req.params;
     const product = await ProductServices.getSingleProductFromDB(productId);
+    if (product === null) {
+      return res.status(404).send({
+        success: false,
+        message: 'Product not found',
+      });
+    }
     res.status(200).send({
       success: true,
       message: 'Product fetched successfully',
@@ -61,6 +76,13 @@ const getSingleProduct = async (req: Request, res: Response) => {
 const updateSingleProduct = async (req: Request, res: Response) => {
   try {
     const { productId } = req.params;
+
+    if (!(await ModelProduct.doesProductExists(productId))) {
+      return res.status(404).send({
+        success: false,
+        message: 'Product not found',
+      });
+    }
     const productData = req.body;
     const result = await ProductServices.updateSingleProductInDB(
       productId,
@@ -72,6 +94,7 @@ const updateSingleProduct = async (req: Request, res: Response) => {
       message: 'Product updated successfully',
       data: result,
     });
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
   } catch (error) {
     res.status(500).send({
       success: false,
@@ -84,11 +107,16 @@ const updateSingleProduct = async (req: Request, res: Response) => {
 const deleteProduct = async (req: Request, res: Response) => {
   try {
     const { productId } = req.params;
-    const result = await ProductServices.deleteProductFromDB(productId);
+    if (!(await ModelProduct.doesProductExists(productId))) {
+      return res.status(404).send({
+        success: false,
+        message: 'Product not found',
+      });
+    }
+    await ProductServices.deleteProductFromDB(productId);
     res.status(200).send({
       success: true,
       message: 'Product deleted successfully',
-      data: result,
     });
   } catch (error) {
     res.status(500).send({
